+++ src/export/ExportMultiple.cpp
@@ -535,14 +535,14 @@ void ExportMultiple::OnExport(wxCommandEvent& WXUNUSED(event))
    {
       wxString msg;
       msg.Printf(
-         ok == eProgressSuccess ? _("Successfully exported the following %ld file(s).")
-           : (ok == eProgressFailed ? _("Something went wrong after exporting the following %ld file(s).")
-             : (ok == eProgressCancelled ? _("Export canceled after exporting the following %ld file(s).")
-               : (ok == eProgressStopped ? _("Export stopped after exporting the following %ld file(s).")
-                 : _("Something went really wrong after exporting the following %ld file(s).")
+         ok == eProgressSuccess ? _("Successfully exported the following %lld file(s).")
+           : (ok == eProgressFailed ? _("Something went wrong after exporting the following %lld file(s).")
+             : (ok == eProgressCancelled ? _("Export canceled after exporting the following %lld file(s).")
+               : (ok == eProgressStopped ? _("Export stopped after exporting the following %lld file(s).")
+                 : _("Something went really wrong after exporting the following %lld file(s).")
                  )
                )
-             ), mExported.GetCount());
+             ), (long long) mExported.GetCount());
 
       wxString FileList;
       for (size_t i = 0; i < mExported.GetCount(); i++) {
